printer/requirements.txt
barcode==1.3
brother_ql==1.1.2
Pillow==9.0.0
